edited Employment, FinalProject, Individuals, Role, Token services an…

+++ SyudentAccounting.BusinessLogic/Services/Implementations/EmploymentService.cs
@@ -2,16 +2,20 @@
 using StudentAccountin.Model.DatabaseModels;
 using StudentAccounting.BusinessLogic.Services.Contracts;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace StudentAccounting.BusinessLogic.Services.Implementations
 {
     public class EmploymentService : IEmploymentService
     {
+        private readonly ILogger<EmploymentService> _logger;
         private readonly ApplicationDatabaseContext _context;
-        public EmploymentService(ApplicationDatabaseContext context)
+        public EmploymentService(ApplicationDatabaseContext context, ILogger<EmploymentService> logger)
         {
+            _logger = logger;
             _context = context;
         }
+        
         public void Create(Employment employment)
         {
             try
@@ -21,42 +25,71 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
             }
             catch (DbUpdateException ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
             }
         }
+        
         public IEnumerable<Employment> Get()
         {
             try
             {
-                return _context.Employments.Include(x => x.Position.Department).Include(x => x.Participants).Include(x=> x.Participants.Individuals).Include(x => x.FinalProjects).AsNoTracking().ToList();
+                var employments = _context.Employments.Include(x => x.Position.Department).
+                    Include(x => x.Participants).Include(x=> x.Participants.Individuals).Include(x => x.FinalProjects).AsNoTracking().ToList();
+
+                return employments;
             }
             catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
+
+                return new List<Employment>();
             }
         }
+        
         public Employment Get(int id)
         {
             try
             {
-                return _context.Employments.Include(x => x.Position.Department).Include(x => x.Participants).Include(x => x.FinalProjects).AsNoTracking().FirstOrDefault(x => x.Id == id);
+                var employment = _context.Employments.Include(x => x.Position.Department).
+                    Include(x => x.Participants).Include(x => x.FinalProjects).AsNoTracking().FirstOrDefault(x => x.Id == id);
+
+                if (employment == null)
+                {
+                    return new Employment();
+                }
+                
+                return employment;
             }
             catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
+
+                return new Employment();
             }
         }
+        
         public Employment GetByParticipants(int participantsId)
         {
             try
             {
-                return _context.Employments.AsNoTracking().Include(x=> x.Position).ThenInclude(x=> x.Department).FirstOrDefault(x => x.ParticipantsId == participantsId);
+                var employment = _context.Employments.AsNoTracking().Include(x=> x.Position).
+                    ThenInclude(x=> x.Department).FirstOrDefault(x => x.ParticipantsId == participantsId);
+
+                if (employment == null)
+                {
+                    return new Employment();
+                }
+                
+                return employment;
             }
-            catch (DbUpdateException ex)
+            catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
+
+                return new Employment();
             }
         }
+        
         public void Edit(Employment employment)
         {
             try
@@ -64,22 +97,27 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
                 _context.Employments.Update(employment);
                 _context.SaveChanges();
             }
-            catch (DbUpdateException ex)
+            catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
             }
         }
+        
         public void Delete(int id)
         {
             try
             {
                 var employment = _context.Employments.FirstOrDefault(x => x.Id == id);
-                _context.Employments.Remove(employment);
-                _context.SaveChanges();
+
+                if (employment != null)
+                {
+                    _context.Employments.Remove(employment);
+                    _context.SaveChanges();
+                }
             }
-            catch (DbUpdateException ex)
+            catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
             }
         }
     }
